+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+  * Prevent Transport::postInstall() from clearing session settings
+
 ### Teleport 1.5.0 (2016-03-01)
 
   * Make sure xPDOObjectCollection sets are ordered